% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis_to_db.R
\name{add_efast_sim_results_from_csv_files}
\alias{add_efast_sim_results_from_csv_files}
\title{Adds a set of results to the database for a given eFAST experiment, from a set of CSV files}
\usage{
add_efast_sim_results_from_csv_files(dblink, results_folder_path, parameters,
  measures, num_curves, experiment_id = NULL, experiment_description = NULL,
  experiment_date = NULL)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{results_folder_path}{The path to folder containing the CSV files being added to the database}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}

\item{num_curves}{Number of resample curves employed in sampling}

\item{experiment_id}{Experiment ID for the results being added. May be NULL if description and date specified}

\item{experiment_description}{A description of this experiment. May be NULL if adding by experiment ID}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}
}
\description{
Adds a set of results to the database for a given eFAST experiment, from a set of CSV files
}
