% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulated_sa_to_db.R
\name{add_emulated_lhc_to_db}
\alias{add_emulated_lhc_to_db}
\title{Takes an emulated LHC experiment and adds the parameter sets and results to the database}
\usage{
add_emulated_lhc_to_db(dblink, parameter_value_set, emulation_predictions,
  prcc_vals, experiment_id = NULL, experiment_description = NULL,
  experiment_date = Sys.Date())
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameter_value_set}{Set of parameters used in the experiment}

\item{emulation_predictions}{Emulation predictions for this set of parameters}

\item{prcc_vals}{Calculated PRCC values for this experiment}

\item{experiment_id}{Experiment ID for the results being added. May be NULL if description and date specified}

\item{experiment_description}{A description of this experiment. May be NULL if adding by experiment ID}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}
}
\description{
Takes an emulated LHC experiment and adds the parameter sets and results to the database
}
