% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_generation.R
\name{add_emulator_performance_stats_to_db}
\alias{add_emulator_performance_stats_to_db}
\title{Adds emulator performance stats to the database for retrieval later}
\usage{
add_emulator_performance_stats_to_db(dblink, built_emulators, measures,
  new_experiment_id)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{built_emulators}{Object containing the emulators built by spartan}

\item{measures}{Simulation output responses}

\item{new_experiment_id}{ID of the emulator generation experiment in the database}
}
\description{
Adds emulator performance stats to the database for retrieval later
}
