% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_generation.R
\name{add_ensemble_performance_stats_to_db}
\alias{add_ensemble_performance_stats_to_db}
\title{Add performance statistics for the ensemble to the database}
\usage{
add_ensemble_performance_stats_to_db(dblink, generated_ensemble, measures,
  new_experiment_id)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{generated_ensemble}{Object containing the ensemble and emulators built by spartan}

\item{measures}{Simulation output responses}

\item{new_experiment_id}{ID of the emulator generation experiment in the database}
}
\description{
Add performance statistics for the ensemble to the database
}
