% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{add_existing_efast_sample_to_database}
\alias{add_existing_efast_sample_to_database}
\title{Adds a previously generated eFAST sample to the database, either creating an experiment record for this in doing so or checking a specified experiment ID}
\usage{
add_existing_efast_sample_to_database(dblink, parameters, num_curves,
  parameter_set_path = NULL, parameters_r_object = NULL,
  experiment_id = NULL, experiment_description = NULL,
  experiment_date = Sys.Date())
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{Simulation parameters being analysed}

\item{num_curves}{Number of resample curves employed in sampling}

\item{parameter_set_path}{Path to the parameter sets to add to the database, if in CSV files}

\item{parameters_r_object}{Name of R object in environment, in which samples values reside}

\item{experiment_id}{The ID of the experiment in the database, if not a new
experiment. If NULL an experiment ID will be created}

\item{experiment_description}{A description of this experiment, if a new
experiment is being created}

\item{experiment_date}{Date experiment created. Defaults to today's date
if not entered}
}
\description{
If an experiment ID is specified, the existance of this record is checked, as well as the existance of any current parameter sets for that experiment. In
this case, as spartan creates an eFAST sample over several files, the user should provide the folder containing each of these files
}
