% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis_to_db.R
\name{add_lhc_and_robustness_sim_results_from_csv_file}
\alias{add_lhc_and_robustness_sim_results_from_csv_file}
\title{Adds a set of results to the database for a given LHC or robustness experiment, from a CSV file}
\usage{
add_lhc_and_robustness_sim_results_from_csv_file(dblink, all_results,
  parameters, measures, experiment_type, experiment_id = NULL,
  experiment_description = NULL, experiment_date = NULL)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{all_results}{The path to the CSV file being added to the database}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}

\item{experiment_type}{Whether this is an LHC or Robustness analysis experiment}

\item{experiment_id}{Experiment ID for the results being added. May be NULL if description and date specified}

\item{experiment_description}{A description of this experiment. May be NULL if adding by experiment ID}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}
}
\description{
Adds a set of results to the database for a given LHC or robustness experiment, from a CSV file
}
