% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_db_summary_responses.R
\name{add_prcc_values_to_db}
\alias{add_prcc_values_to_db}
\title{Adds a set of PRCC values for an LHC experiment to the database.}
\usage{
add_prcc_values_to_db(dblink, parameters, measures, coeffs, experiment_id)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{Simulation parameters being examined}

\item{measures}{Simulation output responses}

\item{coeffs}{Calculated values of PRCC (using spartan method)}

\item{experiment_id}{Experiment ID for the results being added.}
}
\description{
Adds a set of PRCC values for an LHC experiment to the database.
}
