% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_generation.R
\name{build_emulators_from_db_data}
\alias{build_emulators_from_db_data}
\title{Takes a partitioned data set and uses spartan to generate the requested emulations from that data}
\usage{
build_emulators_from_db_data(emulator_list, partitioned_data, parameters,
  measures, normalise_set)
}
\arguments{
\item{emulator_list}{Vector of the types of emulation model to create.
Accepted abbreviations are: SVM (Support-Vector Machine), GP (Gaussian
Process Model), NNET (Neural Network), RF (Random Forest), GLM (General
Linear Model)}

\item{partitioned_data}{Partitioned data object returned from spartan}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}

\item{normalise_set}{Whether the data has been normalised}
}
\value{
Object containing the built emulators
}
\description{
Takes a partitioned data set and uses spartan to generate the requested emulations from that data
}
