% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis_to_db.R
\name{check_experiment_and_parameters_exist_for_adding_results}
\alias{check_experiment_and_parameters_exist_for_adding_results}
\title{When adding results to the database, checks that an experiment exists in the database with associated parameter value sets}
\usage{
check_experiment_and_parameters_exist_for_adding_results(dblink,
  experiment_id = NULL, experiment_description = NULL,
  experiment_date = NULL)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{experiment_id}{Experiment ID for the results being added. May be NULL if description and date specified}

\item{experiment_description}{A description of this experiment. May be NULL if adding by experiment ID}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}
}
\value{
Either the ID of the experiment or a -1 if either the experiment does not exist in the DB, or the experiment has no associated parameters
}
\description{
When adding results to the database, checks that an experiment exists in the database with associated parameter value sets
}
