% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{check_experiment_id}
\alias{check_experiment_id}
\title{Creates a new experiment ID if one not specified, and checks ID is in database with no parameter sets if one is specified}
\usage{
check_experiment_id(dblink, experiment_id, experiment_type, experiment_date,
  experiment_description)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{experiment_id}{Experiment ID for parameter set being generated. May be NULL to create one, or specified}

\item{experiment_type}{Type of spartan sample being generated}

\item{experiment_date}{Date experiment created.}

\item{experiment_description}{A description of this experiment, if a new
experiment is being created}
}
\value{
Either a new experiment ID, the specified experiment ID if no error, or -1 if error encountered
}
\description{
Creates a new experiment ID if one not specified, and checks ID is in database with no parameter sets if one is specified
}
