% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utilities.R
\name{check_parameter_sets_exist_for_given_experiment_id}
\alias{check_parameter_sets_exist_for_given_experiment_id}
\title{Determines whether parameters have been added for a given experiment ID}
\usage{
check_parameter_sets_exist_for_given_experiment_id(dblink, experiment_id)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{experiment_id}{Id of experiment for which parameter sets are being sought}
}
\value{
Boolean showing whether parameter sets exist in database or not
}
\description{
Determines whether parameters have been added for a given experiment ID
}
