% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_db_summary_responses.R
\name{construct_range_info_vectors}
\alias{construct_range_info_vectors}
\title{To save having to input min, max, and inc values for the analysis, this works them out from sampling}
\usage{
construct_range_info_vectors(parameters, db_result)
}
\arguments{
\item{parameters}{Simulation parameters being examined}

\item{db_result}{results of the DB query for parameter sets}
}
\value{
List containing vector for min, max, and inc values used in sampling
}
\description{
To save having to input min, max, and inc values for the analysis, this works them out from sampling
}
