% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_setup.R
\name{create_analysed_results_table}
\alias{create_analysed_results_table}
\title{Creates database table that stores analysed results for all parameter sets for each experiment}
\usage{
create_analysed_results_table(dblink, measures)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{measures}{The measures of the simulation that are being assessed}
}
\description{
The results table consists of results of replicate executions of the simulation under the same
parameter set. This table contains analysed results, or summarises, derived from these replicates.
So for example, in an LHC, each parameter set is executed a number of times, and all responses
stored in the results table. These are then summarised to calculate a median response for those
distributions, and those summary statistics stored in this table.
}
