% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_setup.R
\name{create_database_structure}
\alias{create_database_structure}
\title{Creates all required databases for storing experiments using a simulator}
\usage{
create_database_structure(dblink, parameters, measures)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}
}
\description{
Takes a connection to a mysql database and creates all the tables required
for storing results of all experiments conducted using a simulator
}
