% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_generation.R
\name{create_emulators_from_database_experiments}
\alias{create_emulators_from_database_experiments}
\title{Takes a list of database experiment ID's  and creates emulators from that stored data}
\usage{
create_emulators_from_database_experiments(dblink, parameters, measures,
  emulator_list, percent_train = 75, percent_test = 15,
  percent_validation = 10, normalise_set = FALSE, experiment_id)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}

\item{emulator_list}{Vector of the types of emulation model to create.
Accepted abbreviations are: SVM (Support-Vector Machine), GP (Gaussian
Process Model), NNET (Neural Network), RF (Random Forest), GLM (General
Linear Model)}

\item{percent_train}{Percent of the dataset to use as training}

\item{percent_test}{Percent of the dataset to use as testing}

\item{percent_validation}{Percent of the dataset to use as validation}

\item{normalise_set}{Whether data retrieved from the database should be normalised}

\item{experiment_id}{List of experiment IDs in the database that form the test, training, and validation sets}
}
\description{
Takes a list of database experiment ID's  and creates emulators from that stored data
}
