% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_setup.R
\name{create_experiments_table}
\alias{create_experiments_table}
\title{Creates database table that stores link to each experiment}
\usage{
create_experiments_table(dblink)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}
}
\description{
Creates a table that gives each experiment its' own ID, stores the type
of analysis performed (LHC, Robustness, eFAST), the date entered and a
text description of that study.
}
