% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_setup.R
\name{create_field_string}
\alias{create_field_string}
\title{Creates the mySQL string to create columns for each parameter, or measure, required in the table}
\usage{
create_field_string(field_list, add_analysis_specific_fields = TRUE)
}
\arguments{
\item{field_list}{List of columns required in the database. Could be parameters or simulation
response measures.}

\item{add_analysis_specific_fields}{Whether paramOfInterest and curve should be added. Not added in
all cases of table creation}
}
\value{
String to add to a mySQL query when creating a table
}
\description{
Creates the mySQL string to create columns for each parameter, or measure, required in the table
}
