% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_setup.R
\name{create_parameter_values_table}
\alias{create_parameter_values_table}
\title{Creates database table that stores all generated parameter sets for each experiment}
\usage{
create_parameter_values_table(dblink, parameters)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{The parameters of the simulation that are being analysed}
}
\description{
Creates a table that gives each parameter set its' own ID, stores the values for each
of the parameters, and the ID of the experiment to which this set belongs. To ensure
all sampling analyses in spartan can be stored in this database, additional columns for
parameter of interest and curve are added for robustness and eFAST sampling techniques,
yet these remain empty for lhc
}
