% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_setup.R
\name{create_simulation_results_table}
\alias{create_simulation_results_table}
\title{Creates database table that stores results for all parameter sets for each experiment}
\usage{
create_simulation_results_table(dblink, measures)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{measures}{The measures of the simulation that are being assessed}
}
\description{
Creates a table that gives each result set its' own ID, stores the outputs obtained in
simulation for those values, and the parameter and experiment ID to which these
results belong. There may be multiple results for each parameter set in this table,
to account for execution of replicates required to reduce aleatory uncertainty (as
explained in the spartan package)
}
