% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_setup.R
\name{create_stats_table}
\alias{create_stats_table}
\title{Adds a table for the spartan generated statistics for each experiment, such as A-Test scores and PRCCs}
\usage{
create_stats_table(dblink, parameters, measures)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{The parameters of the simulation being assessed}

\item{measures}{The measures of the simulation that are being assessed}
}
\description{
Adds a table for the spartan generated statistics for each experiment, such as A-Test scores and PRCCs
}
