% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_deletion.R
\name{delete_database_structure}
\alias{delete_database_structure}
\title{Remove all tables from the database}
\usage{
delete_database_structure(dblink)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}
}
\description{
Note this function should be used rather than the individual remove table functions
as the order of deletion is set by foreign key usage
}
