% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_deletion.R
\name{delete_parameters_table}
\alias{delete_parameters_table}
\title{Remove the parameters table from the database, if this exists}
\usage{
delete_parameters_table(dblink)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}
}
\description{
Remove the parameters table from the database, if this exists
}
