% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_deletion.R
\name{delete_results_table}
\alias{delete_results_table}
\title{Remove the results table from the database, if this exists}
\usage{
delete_results_table(dblink)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}
}
\description{
Remove the results table from the database, if this exists
}
