% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_table_deletion.R
\name{delete_statistics_table}
\alias{delete_statistics_table}
\title{Remove the statistics table from the database, if this exists}
\usage{
delete_statistics_table(dblink)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}
}
\description{
Remove the statistics table from the database, if this exists
}
