% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{download_sample_as_csvfile}
\alias{download_sample_as_csvfile}
\title{Download a generated sample from the database as a CSV file}
\usage{
download_sample_as_csvfile(output_dir, dblink, experiment_type,
  experiment_id = NULL, experiment_description = NULL,
  experiment_date = NULL)
}
\arguments{
\item{output_dir}{Directory to store the generated CSV file}

\item{dblink}{A link to the database in which this entry is being added}

\item{experiment_type}{Type of sample generated (Robustness, LHC, eFAST)}

\item{experiment_id}{Id of the experiment in the database}

\item{experiment_description}{Description of the experiment being performed}

\item{experiment_date}{Date the experiment was performed}
}
\description{
Download a generated sample from the database as a CSV file
}
