% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_db_summary_responses.R
\name{generate_efast_analysis}
\alias{generate_efast_analysis}
\title{From the results stored in a mysql database, produce summary statistics using eFAST technique}
\usage{
generate_efast_analysis(dblink, parameters, measures, experiment_id = NULL,
  experiment_date = NULL, experiment_description = NULL)
}
\arguments{
\item{dblink}{A link to the database in which the results are stored}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}

\item{experiment_id}{Experiment ID for the experiment being plotted. May be NULL if description and date specified}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}

\item{experiment_description}{A description of this experiment. May be NULL if adding by experiment ID}
}
\description{
From the results stored in a mysql database, produce summary statistics using eFAST technique
}
