% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{generate_efast_set_in_db}
\alias{generate_efast_set_in_db}
\title{Generate eFAST analysis sample and store in database}
\usage{
generate_efast_set_in_db(dblink, parameters, num_samples, minvals, maxvals,
  num_curves, experiment_id = NULL, experiment_description = NULL)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{num_samples}{The number of parameter subsets to generate - should be at
least 65 for eFAST}

\item{minvals}{The minimum value of the range to explore for each parameter}

\item{maxvals}{The maximum value of the range to explore for each parameter}

\item{num_curves}{The number of 'resamples' to perform (see eFAST
documentation) - recommend using at least 3}

\item{experiment_id}{The ID of the experiment in the database, if not a new
experiment. If NULL an experiment ID will be created}

\item{experiment_description}{A description of this experiment, if a new
experiment is being created}
}
\description{
Generates a set of parameters for execution in simulation to permit
an eFAST analyis. These parameters are stored in the database,
either for a given experiment ID created previously, or a new experiment
is created in the database
}
