% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_db_summary_responses.R
\name{generate_lhc_analysis}
\alias{generate_lhc_analysis}
\title{Using results stored in the database, calculates partial rank correlation coefficients for
all measure-parameter pairs in an LHC experiment}
\usage{
generate_lhc_analysis(dblink, parameters, measures, experiment_id = NULL,
  experiment_date = NULL, experiment_description = NULL)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{Simulation parameters being examined}

\item{measures}{Simulation output responses}

\item{experiment_id}{Experiment ID for the results being added. May be NULL if description and date specified}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}

\item{experiment_description}{A description of this experiment. May be NULL if adding by experiment ID}
}
\description{
Using results stored in the database, calculates partial rank correlation coefficients for
all measure-parameter pairs in an LHC experiment
}
