% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{generate_lhc_set_in_db}
\alias{generate_lhc_set_in_db}
\title{Generate LHC analysis sample and store in database}
\usage{
generate_lhc_set_in_db(dblink, parameters, num_samples, minvals, maxvals,
  algorithm, experiment_id = NULL, experiment_description = NULL)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{num_samples}{The number of samples to take from the hypercube}

\item{minvals}{The minimum value of the range to explore for each parameter}

\item{maxvals}{The maximum value of the range to explore for each parameter}

\item{algorithm}{Choice of algorithm to use to generate the hypercube.
Can be set to either 'normal' or 'optimum'. Beware optimum can take a
long time to generate an optimised parameter set (more than 24 hours
in some circumstances)}

\item{experiment_id}{The ID of the experiment in the database, if not a new
experiment. If NULL an experiment ID will be created}

\item{experiment_description}{A description of this experiment, if a new
experiment is being created}
}
\description{
Generates a set of parameters for execution in simulation to permit
an LHC/PRCC analysis as described in spartan. These parameters are stored in the database,
either for a given experiment ID created previously, or a new experiment
is created in the database
}
