% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{generate_robustness_set_in_db}
\alias{generate_robustness_set_in_db}
\title{Generate robustness analysis sample and store in database}
\usage{
generate_robustness_set_in_db(dblink, parameters, baseline, minvals, maxvals,
  incvals, experiment_id = NULL, experiment_description = NULL,
  experiment_date = Sys.Date())
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{baseline}{The calibrated value of each parameter being analysed}

\item{minvals}{The minimum value of the range to explore for each parameter}

\item{maxvals}{The maximum value of the range to explore for each parameter}

\item{incvals}{The sampling increment value to apply to each parameter between
the minimum and maximum limits}

\item{experiment_id}{The ID of the experiment in the database, if not a new
experiment. If NULL an experiment ID will be created}

\item{experiment_description}{A description of this experiment, if a new
experiment is being created}

\item{experiment_date}{Date experiment created. Defaults to today's date
if not entered}
}
\description{
Generates a set of parameters for execution in simulation to permit
a robustness analysis. These parameters are stored in the database,
either for a given experiment ID created previously, or a new experiment
is created in the database
}
