% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utilities.R
\name{get_experiment_id}
\alias{get_experiment_id}
\title{Return the experiment ID for a given experiment type, date, and description}
\usage{
get_experiment_id(dblink, experiment_type, experiment_date,
  experiment_description)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{experiment_type}{Whether the experiment was a consistency, robustness, LHC, or eFAST analysis}

\item{experiment_date}{Date the experiment was performed}

\item{experiment_description}{Description of the experiment being performed}
}
\value{
Integer value for the ID of this experiment in the database
}
\description{
Return the experiment ID for a given experiment type, date, and description
}
