% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_db_summary_responses.R
\name{graph_efast_analysis}
\alias{graph_efast_analysis}
\title{Mine the DB for an eFAST experiment and plot the results}
\usage{
graph_efast_analysis(dblink, parameters, measures, output_directory,
  experiment_id = NULL, experiment_date = NULL,
  experiment_description = NULL)
}
\arguments{
\item{dblink}{A link to the database in which the results are stored}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}

\item{output_directory}{Where graphs should be output to}

\item{experiment_id}{Experiment ID for the experiment being plotted. May be NULL if description and date specified}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}

\item{experiment_description}{A description of this experiment. May be NULL if adding by experiment ID}
}
\description{
Mine the DB for an eFAST experiment and plot the results
}
