% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_db_summary_responses.R
\name{graph_lhc_analysis}
\alias{graph_lhc_analysis}
\title{Take LHC experiment from database and produce plots of the results}
\usage{
graph_lhc_analysis(dblink, parameters, measures, measure_scale,
  output_directory, experiment_id = NULL, experiment_date = NULL,
  experiment_description = NULL, output_type = c("PDF"))
}
\arguments{
\item{dblink}{A link to the database in which the results are stored}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}

\item{measure_scale}{The scale in which each response is measured. Included on the output plot}

\item{output_directory}{Where graphs should be output to}

\item{experiment_id}{Experiment ID for the experiment being plotted. May be NULL if description and date specified}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}

\item{experiment_description}{A description of this experiment. May be NULL if adding by experiment ID}

\item{output_type}{Type of graph to produce. Can be PDF, TIFF, PNG, BMP}
}
\description{
Take LHC experiment from database and produce plots of the results
}
