% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_db_summary_responses.R
\name{graph_robustness_analysis}
\alias{graph_robustness_analysis}
\title{Function to retrieve robustness results from the database and produce plots}
\usage{
graph_robustness_analysis(dblink, output_directory, parameters, measures,
  ATESTSIGLEVEL = 0.23, experiment_id = NULL, experiment_date = NULL,
  experiment_description = NULL)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{output_directory}{Where the graphs should be stored}

\item{parameters}{Simulation parameters being examined}

\item{measures}{Simulation output responses}

\item{ATESTSIGLEVEL}{Level above 0.5 at which a difference is deemed a large difference}

\item{experiment_id}{Experiment ID for the results being added. May be NULL if description and date specified}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}

\item{experiment_description}{A description of this experiment. May be NULL if adding by experiment ID}
}
\description{
Function to retrieve robustness results from the database and produce plots
}
