% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ppsim_robustness_set}
\alias{ppsim_robustness_set}
\title{Robustness example parameter sample file}
\format{A list with 80 rows (one per parameter set) and seven columns}
\usage{
data(ppsim_robustness_set)
}
\description{
This is a dataset containing values for the six parameters that control the
simulation detailed in the case study section of the spartan vignette, with
sampling performed for a robustness analysis. These parameters are defined
in the accompanying publications referred to in the vignette
}
\details{
\itemize{
  \item stableBindProbability. Parameter values between 0 and 100
  \item chemokineExpressionThreshold. Parameter values between 0 and 1
  \item initialChemokineExpressionValue. Parameter values between 0.1 and 0.5
  \item maxChemokineExpressionValue. Parameter values between 0.015 and 0.08
  \item maxProbabilityOfAdhesion. Parameter values between 0 and 1
  \item adhesionFactorExpressionSlope. Parameter values between 0.25 and 5
  \item Velocity. Simulation output for cell velocity
  \item Displacment. Simulation output for cell displacement
  \item paramOfInterest. Which parameter is being considered in this robustness analysis
}
}
\keyword{datasets}
