% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{pregenerated_lhc}
\alias{pregenerated_lhc}
\title{Latin-hypercube parameter value set to demonstrate storage of pre-generated sample file}
\format{A list with 500 rows (one per parameter set) and six columns}
\usage{
data(pregenerated_lhc)
}
\description{
This is a dataset containing values for the six parameters that control the
simulation detailed in the case study section of the spartan vignette. These
parameters are defined the accompanying publications referred to in
the vignette
}
\details{
\itemize{
  \item stableBindProbability. Parameter values between 0 and 100
  \item chemokineExpressionThreshold. Parameter values between 0 and 1
  \item initialChemokineExpressionValue. Parameter values between 0.1 and 0.5
  \item maxChemokineExpressionValue. Parameter values between 0.015 and 0.08
  \item maxProbabilityOfAdhesion. Parameter values between 0 and 1
  \item adhesionFactorExpressionSlope. Parameter values between 0.25 and 5
}
}
\keyword{datasets}
