% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_generation.R
\name{record_partitioned_data_measures_to_db}
\alias{record_partitioned_data_measures_to_db}
\title{Takes the partitioned output measures values data and stores this in the database, so emulation creation can be replicated if necessary}
\usage{
record_partitioned_data_measures_to_db(dblink, partitioned_data, measures,
  new_experiment_id)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{partitioned_data}{Partitioned dataset being used in emulator generation, to add to DB}

\item{measures}{Simulation output responses}

\item{new_experiment_id}{ID for this experiment in the database}
}
\description{
In doing this, whether the row of the data is training, testing, or validation is stored in the paramOfInterest column
}
