% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_generation.R
\name{regenerate_emulators_from_db_data}
\alias{regenerate_emulators_from_db_data}
\title{Takes a previous emulator generation experiment stored in the database and regenerates the emulators. Ensures emulators can be regenerated from the same data}
\usage{
regenerate_emulators_from_db_data(dblink, parameters, measures,
  emulator_list, normalise_set = FALSE, experiment_id = NULL,
  experiment_description = NULL, experiment_date = NULL)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}

\item{emulator_list}{Vector of the types of emulation model to create.
Accepted abbreviations are: SVM (Support-Vector Machine), GP (Gaussian
Process Model), NNET (Neural Network), RF (Random Forest), GLM (General
Linear Model)}

\item{normalise_set}{Whether data retrieved from the database should be normalised}

\item{experiment_id}{List of experiment IDs in the database that form the test, training, and validation sets}

\item{experiment_description}{If not searching for previous result by ID, can use experiment description and date instead}

\item{experiment_date}{If not searching for previous result by ID, can use experiment description and date instead}
}
\description{
Takes a previous emulator generation experiment stored in the database and regenerates the emulators. Ensures emulators can be regenerated from the same data
}
