% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{remove_errored_new_experiment_from_db}
\alias{remove_errored_new_experiment_from_db}
\title{Used when creating a new experiment for a new sample, yet an error is encountered in sampling. Removes generated experiment}
\usage{
remove_errored_new_experiment_from_db(dblink, experiment_id)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{experiment_id}{Id of the experiment created that needs to be deleted}
}
\description{
This ensures no new experiments exist in the database where an error has been encountered in sampling. If the user specified
an experiment ID to use when generating a sample, that experiment ID is not removed
}
