% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulated_abc_to_db.R
\name{retrieve_abc_experiment_for_plotting}
\alias{retrieve_abc_experiment_for_plotting}
\title{Retrieve posteriors from database and produce density plots}
\usage{
retrieve_abc_experiment_for_plotting(dblink, parameters,
  experiment_id = NULL, experiment_description = NULL,
  experiment_date = Sys.Date())
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{Simulation parameters being examined}

\item{experiment_id}{Experiment ID for the ABC results. May be NULL if description and date specified}

\item{experiment_description}{A description of this ABC experiment. May be NULL if adding by experiment ID}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}
}
\description{
Retrieve posteriors from database and produce density plots
}
