% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utilities.R
\name{retrieve_parameter_of_interest}
\alias{retrieve_parameter_of_interest}
\title{For a given parameter set in the database, retrieve the parameter of interest for that set in the analysis}
\usage{
retrieve_parameter_of_interest(dblink, parameter_set, experiment_id)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameter_set}{Parameter set for which ID is being sought}

\item{experiment_id}{Experiment ID for which this parameter set is within}
}
\value{
Parameter of interest in this analysis
}
\description{
For a given parameter set in the database, retrieve the parameter of interest for that set in the analysis
}
