% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_generation.R
\name{retrieve_validation_set_from_db_for_emulator}
\alias{retrieve_validation_set_from_db_for_emulator}
\title{To demonstrate use of emulator, we return validation set from the database to show how the emulator can make predictions on output}
\usage{
retrieve_validation_set_from_db_for_emulator(dblink, parameters, measures,
  experiment_id)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}

\item{experiment_id}{ID of the experiment for which the validation set is being returned}
}
\value{
Set of parameter values that can be used to make predictions
}
\description{
To demonstrate use of emulator, we return validation set from the database to show how the emulator can make predictions on output
}
