% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{set_new_experiment_flag}
\alias{set_new_experiment_flag}
\title{Sets the flag used to denote whether a new experiment was created in sampling. Used to roll DB back on error}
\usage{
set_new_experiment_flag(experiment_id)
}
\arguments{
\item{experiment_id}{Specified experiment ID to function. If NULL, a new experiment was created}
}
\value{
Boolean flag showing whether a new experiment ID is needed (TRUE) or one was specified (FALSE)
}
\description{
Sets the flag used to denote whether a new experiment was created in sampling. Used to roll DB back on error
}
