% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{setup_experiment}
\alias{setup_experiment}
\title{Setup a new experiment in the database, returning the generated ID of the experiment}
\usage{
setup_experiment(dblink, experiment_type, experiment_date,
  experiment_description)
}
\arguments{
\item{dblink}{A link to the database in which this entry is being added}

\item{experiment_type}{Whether the experiment was a consistency, robustness, LHC, or eFAST analysis}

\item{experiment_date}{Date the experiment was performed}

\item{experiment_description}{Description of the experiment being performed}
}
\value{
Generated ID for this experiment (the database table primary key)
}
\description{
Setup a new experiment in the database, returning the generated ID of the experiment
}
