% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulated_abc_to_db.R
\name{store_abc_experiment_in_db}
\alias{store_abc_experiment_in_db}
\title{Store an ABC experiment in the simulation database}
\usage{
store_abc_experiment_in_db(dblink, abc_settings, abc_results, parameters,
  experiment_id = NULL, experiment_description = NULL,
  experiment_date = Sys.Date(), graph_results = FALSE,
  output_directory = NULL)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{abc_settings}{Settings object created when running the ABC analysis}

\item{abc_results}{Object produced by the EasyABC ABC_Sequential method, containing analysis results}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{experiment_id}{Experiment ID for the results being added. May be NULL if description and date specified}

\item{experiment_description}{A description of this experiment. May be NULL if adding by experiment ID}

\item{experiment_date}{Date experiment created. May be NULL if adding by experiment ID}

\item{graph_results}{Whether the graphs should be plotted}

\item{output_directory}{If graph_results is TRUE, where the graph should be plotted to}
}
\description{
Store an ABC experiment in the simulation database
}
