% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utilities.R
\name{store_summary_experiment_result}
\alias{store_summary_experiment_result}
\title{Provides a means for storing the results of an experiment in the database. This assumes sets where the execution was done once (not replicate runs of parameter sets)}
\usage{
store_summary_experiment_result(dblink, experiment_parameter_set,
  experiment_result_set, experiment_id = NULL,
  experiment_description = NULL, experiment_date = NULL)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{experiment_parameter_set}{Parameters for this experiment}

\item{experiment_result_set}{Simulation responses under those parameter conditions}

\item{experiment_id}{Either existing ID of experiment for which this result is being added, or NULL if a new experiment}

\item{experiment_description}{Description of the experiment being performed. Can be NULL if using existing experiment ID}

\item{experiment_date}{Date the experiment was performed. Can be NULL if using existing experiment ID}
}
\description{
Provides a means for storing the results of an experiment in the database. This assumes sets where the execution was done once (not replicate runs of parameter sets)
}
