% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_generation.R
\name{use_emulators_to_make_and_store_predictions}
\alias{use_emulators_to_make_and_store_predictions}
\title{Takes a set of parameters and uses a set of emulators to generate predictions of simulator output, then adds these to the database}
\usage{
use_emulators_to_make_and_store_predictions(dblink, sim_emulators,
  parameters, measures, data_to_predict, normalise = FALSE,
  normalise_result = FALSE, experiment_description,
  experiment_date = Sys.Date())
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{sim_emulators}{Emulation object created by spartan}

\item{parameters}{The parameters of the simulation that are being analysed}

\item{measures}{The measures of the simulation that are being assessed}

\item{data_to_predict}{Sets of parameters for which simulation output measures should be predicted}

\item{normalise}{Whether the data_to_predict should be normalised prior to making predictions. Defaults to FALSE}

\item{normalise_result}{Whether the predictions generated should be normalised using the scale set in the emulator. Defaults to FALSE}

\item{experiment_description}{Description of this experiment, to be added to the database}

\item{experiment_date}{Date this experiment was performed. Defaults to that day's date}
}
\description{
Takes a set of parameters and uses a set of emulators to generate predictions of simulator output, then adds these to the database
}
