% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{add_existing_lhc_sample_to_database}
\alias{add_existing_lhc_sample_to_database}
\title{Adds a previously generated LHC sample to the database, either creating an experiment record for this in doing so or checking a specified experiment ID}
\usage{
add_existing_lhc_sample_to_database(dblink, parameter_set,
  experiment_id = NULL, experiment_description = NULL,
  experiment_date = Sys.Date())
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameter_set}{Parameter set to add to the database}

\item{experiment_id}{The ID of the experiment in the database, if not a new
experiment. If NULL an experiment ID will be created}

\item{experiment_description}{A description of this experiment, if a new
experiment is being created}

\item{experiment_date}{Date experiment created. Defaults to today's date
if not entered}
}
\description{
If an experiment ID is specified, the existance of this record is checked, as well as the existance of any current parameter sets for that experiment
}
