% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_sample_in_db.R
\name{add_parameter_set_to_database}
\alias{add_parameter_set_to_database}
\title{Add a generated parameter set to the database}
\usage{
add_parameter_set_to_database(dblink, parameter_set, experiment_id)
}
\arguments{
\item{dblink}{A link to the database in which this table is being created}

\item{parameter_set}{Parameter set to add to the database}

\item{experiment_id}{Database experiment ID to which this parameter set belongs}
}
\description{
Add a generated parameter set to the database
}
